Exchange_Rate
The application I developed is designed to track real-time exchange rates. It displays the current conversion rates between different currencies such as USD-TL, TL-USD, EUR-TL, and TL-EUR. This allows users to instantly view the exchange rates for their desired currency pairs. On the backend, the application fetches exchange rate data from an API to keep the information up-to-date, while the frontend presents this data in a user-friendly interface. This project, built using React and Axios, provides an effective solution for monitoring exchange rates.